+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [CHANGE] Drop support for Ruby 2.5.x (by [@cleicar][])
 * [CHANGE] Drop support for Ruby 2.4.x (by [@rishijain][])
+* [FEATURE] Add ruby_extensions configuration option (by [@stufro][])
+* [CHANGE] Include files which have a ruby shebang (by [@stufro][])
 
 # v4.7.0 / 2022-05-06 [(commits)](https://github.com/whitesmith/rubycritic/compare/v4.6.1...v4.7.0)